Delete commands.go
